backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
Pillow
